split out types & systems from GameWorker
