+++ src/engine/component/renderable.ts
@@ -3,9 +3,9 @@ import { addComponent, hasComponent, IComponent } from "bitecs";
 import { renderableBuffer } from "./buffers";
 import { addView } from "../allocator/CursorBuffer";
 import { maxEntities } from "../config.common";
-import { GameState, World } from "../GameWorker";
 import { SetActiveCameraMessage, SetActiveSceneMessage, WorkerMessageType } from "../WorkerMessage";
 import { traverse } from "./transform";
+import { GameState, World } from "../GameTypes";
 
 export interface Renderable extends IComponent {
   resourceId: Uint32Array;
