+++ src/plugins/PhysicsCharacterController.ts
@@ -5,7 +5,7 @@ import { Object3D, Quaternion, Vector3 } from "three";
 import { Player } from "../engine/component/Player";
 import { addRenderableComponent } from "../engine/component/renderable";
 import { addChild, addTransformComponent, Transform } from "../engine/component/transform";
-import { GameState } from "../engine/GameWorker";
+import { GameState } from "../engine/GameTypes";
 import { ButtonActionState } from "../engine/input/ActionMappingSystem";
 import { InputModule } from "../engine/input/input.game";
 import { getModule } from "../engine/module/module.common";
